+++ BarberShop/Services/Interfaces/IExistingEntityCheckerService.cs
@@ -8,7 +8,7 @@
         Task<bool> IsCorrectBarbershopId(long id, CancellationToken cancellationToken);
         Task<bool> IsCorrectCommentId(long id, CancellationToken cancellationToken);
         Task<bool> IsCorrectEmployeeId(long id, CancellationToken cancellationToken);
-        Task<bool> IsCorrectPositionId(long id, CancellationToken cancellationToken);
+        Task<bool> IsCorrectPositionId(int id, CancellationToken cancellationToken);
         Task<bool> IsCorrectReservationId(long id, CancellationToken cancellationToken);
         Task<bool> IsCorrectServiceId(long id, CancellationToken cancellationToken); 
         Task<bool> IsCorrectUserId(long id, CancellationToken cancellationToken);
